distance wall fade working

+++ src/game/Logic.h
@@ -107,9 +107,9 @@ typedef struct {
   Arena_t* debugArena;
   Player_t player;
 
-  u8 WORLD_HEIGHT;
-  u8 WORLD_TILE_SCALE;
-  u8 ATLAS_TILE_SIZE;
+  f32 WORLD_HEIGHT;
+  f32 WORLD_TILE_SCALE;
+  f32 ATLAS_TILE_SIZE;
 } Logic__State_t;
 
 void Player__Init(Logic__State_t* game);
